changed type of stats variables

+++ Skeleton/Game/Interfaces/ICharacter.cs
@@ -3,9 +3,9 @@
     public interface ICharacter
     {
         bool IsAlive { get; set; }
-        int HealthPoints { get; set; }
-        int AttackPoints { get; set; }
-        int DefencePoints { get; set; }
+        double HealthPoints { get; set; }
+        double AttackPoints { get; set; }
+        double DefencePoints { get; set; }
         double X { get; set; }
         double Y { get; set; }
     }
